// sbsp_frontend/src/types/ActiveCue.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PlaybackStatus } from "./PlaybackStatus";
import type { StateParam } from "./StateParam";

export type ActiveCue = { cueId: string, position: number, duration: number, status: PlaybackStatus, params: StateParam, };